+++ internal/json/json.go
@@ -3,9 +3,6 @@
 // license that can be found in the LICENSE file.
 
 //go:build !jsoniter && !go_json && !(sonic && avx && (linux || windows || darwin) && amd64)
-// +build !jsoniter
-// +build !go_json
-// +build !sonic !avx !linux,!windows,!darwin !amd64
 
 package json
