+++ internal/json/jsoniter.go
@@ -3,7 +3,6 @@
 // license that can be found in the LICENSE file.
 
 //go:build jsoniter
-// +build jsoniter
 
 package json
